pyproject.toml
[project]
name = "startup"
version = "0.1.0"
description = "Startup AI Agent using OpenAI Agents SDK with Gemini API integration"
readme = "README.md"
authors = [
    { name = "Muhammad Qasim", email = "mqasim111786111@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "pydantic>=2.0.0",
    "asyncio-mqtt>=0.16.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.8.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0"
]

[project.scripts]
startup = "startup:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
